+++ test/test_helper.dart
@@ -29,7 +29,7 @@ class WidgetTestBench extends StatelessWidget {
       child: MediaQuery(
         data: mediaQuery,
         child: MaterialApp(
-          theme: theme,
+          theme: ensureDsTokens(theme),
           localizationsDelegates: const [
             AppLocalizations.delegate,
             GlobalMaterialLocalizations.delegate,
@@ -70,7 +70,7 @@ class DarkWidgetTestBench extends StatelessWidget {
       child: MediaQuery(
         data: mediaQuery,
         child: MaterialApp(
-          theme: ThemeData.dark(),
+          theme: ensureDsTokens(ThemeData.dark()),
           localizationsDelegates: const [
             AppLocalizations.delegate,
             GlobalMaterialLocalizations.delegate,
@@ -114,6 +114,7 @@ class RiverpodWidgetTestBench extends StatelessWidget {
       child: MediaQuery(
         data: mediaQuery,
         child: MaterialApp(
+          theme: ensureDsTokens(null),
           localizationsDelegates: const [
             AppLocalizations.delegate,
             GlobalMaterialLocalizations.delegate,
@@ -157,7 +158,7 @@ class DarkRiverpodWidgetTestBench extends StatelessWidget {
       child: MediaQuery(
         data: mediaQuery,
         child: MaterialApp(
-          theme: ThemeData.dark(),
+          theme: ensureDsTokens(ThemeData.dark()),
           localizationsDelegates: const [
             AppLocalizations.delegate,
             GlobalMaterialLocalizations.delegate,